% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sklarsomega.R
\name{logLik.sklarsomega}
\alias{logLik.sklarsomega}
\title{Return the maximum of the Sklar's Omega log objective function.}
\usage{
\method{logLik}{sklarsomega}(object, ...)
}
\arguments{
\item{object}{a fitted model object.}

\item{...}{additional arguments.}
}
\value{
This function returns an object of class \code{logLik}. This is a number with at least one attribute, \code{"df"} (degrees of freedom), giving the number of estimated parameters in the model.
}
\description{
Return the maximum of the Sklar's Omega log objective function.
}
\details{
This function extracts the maximum value of the log objective function from a model fit, and is principally intended to be used in computing information criteria.
}
\seealso{
\code{\link{AIC}}
}

