% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sklarsomega.R
\name{nobs.sklarsomega}
\alias{nobs.sklarsomega}
\title{Return the number of observations for a Sklar's Omega fit.}
\usage{
\method{nobs}{sklarsomega}(object, ...)
}
\arguments{
\item{object}{a fitted model object.}

\item{...}{additional arguments.}
}
\value{
An integer.
}
\description{
Return the number of observations for a Sklar's Omega fit.
}
\details{
This function extracts the number of observations from a model fit, and is principally intended to be used in computing information criteria.
}
\seealso{
\code{\link{AIC}}
}

