% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sklarsomega.R
\name{pcat}
\alias{pcat}
\title{Compute the cumulative distribution function for a categorical distribution.}
\usage{
pcat(q, p)
}
\arguments{
\item{q}{vector of quantiles.}

\item{p}{vector of probabilities.}
}
\value{
\code{pcat} returns a probability or a vector of probabilities, depending on the dimension of \code{q}.
}
\description{
Compute the cumulative distribution function for a categorical distribution.
}
\details{
This function uses \code{\link[LaplacesDemon]{dcat}} to compute the cdf for the categorical distribution with support \eqn{1, \dots , K} and probabilities \eqn{p = (p_1, \dots , p_K)'}.
}

